Updated and disabled diff checker
It was never needed to be used upon a connection, since the diff was never applied to any of the aio hacks to date. It can be un-commented once it is needed, such as hacks that are located in the 0x12000000 range.

Besides that, the checker is updated to support ALL TCPGecko + kernel variants using a pointer, instead of blind guesses of the shifted offsets in the ram.